gmail-extractor/requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
httplib2==0.22.0
idna==3.8
lxml==5.3.0
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
soupsieve==2.6
uritemplate==4.1.1
urllib3==2.2.2